ai_chatbot/backend/requirements.txt
fastapi
uvicorn
langchain
langchain_openai
langchain_community
supabase
python-dotenv
tiktoken
beautifulsoup4
